Allow profiles to store directories
Implement a feature where the list of directories to be scanned is
exteneded with the list in the profile file. The feature is meant for use
cases where the directory is constantly updated.

The syntax in `profile.json` should be as follows.

``````
{
    dir: ["d:\\new folder", "d:\\new folder (2)"]
}
``````

The user will specify "" as the input directory to make use of the stored
directories instead of passing one.