Fix path being passed instead of regex pattern
When specifying a regular expression with the delimiter (e.g. /test/i),
the script infers a relative path and expands it. This is most-likely
default behavior.

Prevent it from happening by escaping the slash (e.g. \/test/i) when in
CLI mode. The bug doesn't affect the pattern in the profile.

In addition, parse the argument with ArgActions.PatternMatchAction and
return a compiled regular expression object.