Fix regression and introduce verbose option
There is a bug which was introduced when changing the indentation to
match PEP8. The final loop which iterates files is outside of the
directory iteration loop. This causes the 'files' array to be overwritten
without the files from being checked. Essentially, this skips the files
in all but the last directory.

Introduce the '--verbose' switch to show detailed information without
overloading the command window when the '--debug' switch has been
specified. Specifying verbose output enable debug as well.